fix!: index bucket output length
The length of the output when unpacking an index bucket was previously
determined by the maximum index value. However, if the bins are bigger
than the data, this can often mean the length of the unpacked groups is
not the same as the number of bins.

The constructor of index bucket has been updated to now also accept the
output dimensions, so that the number of groups and the number of bins
is always the same.